use es6 modules instead of require
