+++ src/input/defaultParams.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
 	accuracy: 0.7,
 	blur: 40,
 	fill: true,
