+++ src/output/node/polygonsToPNGStream.js
@@ -1,8 +1,8 @@
 // https://github.com/Automattic/node-canvas#canvaspngstream
-var Canvas = require( 'canvas-browserify');
-var drawPolygonsOnContext = require('../util/drawPolygonsOnContext.js');
+import Canvas from 'canvas-browserify';
+import drawPolygonsOnContext from '../../util/drawPolygonsOnContext';
 
-module.exports = function ( polygons, size, options ) {
+export default function ( polygons, size, options ) {
 	let dpr = options && options.dpr ? options.dpr : 1;
 	let backgroundColor = options && options.backgroundColor ? options.backgroundColor : false;
 	let canvas = new Canvas( size.width * dpr, size.height * dpr );
