+++ src/polygons/imageDataToPolygons.js
@@ -1,30 +1,31 @@
 'use strict';
+import stackBlur from 'stackblur-canvas';
 
-var stackBlur = require('stackblur-canvas');
-var delaunay = require('delaunay-fast');
+// var stackBlur = require('stackblur-canvas');
+import delaunay from 'delaunay-fast';
 
-var isImageData = require('../util/isImageData.js');
-var copyImageData = require('../imagedata/copyImageData.js');
-var greyscale = require('../imagedata/greyscale');
-var detectEdges = require('../imagedata/detectEdges');
-var getEdgePoints = require('./getEdgePoints.js');
-var getVerticesFromPoints = require('./getVerticesFromPoints.js');
-var addColorToPolygons = require('./addColorToPolygons.js');
+import isImageData from '../util/isImageData';
+import copyImageData from '../imagedata/copyImageData';
+import greyscale from '../imagedata/greyscale';
+import detectEdges from '../imagedata/detectEdges';
+import getEdgePoints from './getEdgePoints';
+import getVerticesFromPoints from './getVerticesFromPoints';
+import addColorToPolygons from './addColorToPolygons';
 
-module.exports = function ( imageData, params ) {
+export default function ( imageData, params ) {
 	
 	if ( isImageData( imageData ) ) {
-		var imageSize = { width: imageData.width, height: imageData.height };
+		let imageSize = { width: imageData.width, height: imageData.height };
 
-		var tmpImageData = copyImageData( imageData );
-		var colorImageData = copyImageData( imageData );
+		let tmpImageData = copyImageData( imageData );
+		let colorImageData = copyImageData( imageData );
 		
-		var blurredImageData = stackBlur.imageDataRGBA( tmpImageData, 0, 0, imageSize.width, imageSize.height, params.blur );
-		var greyscaleImageData = greyscale( blurredImageData );
-		var edgesImageData = detectEdges( greyscaleImageData );
-		var edgePoints = getEdgePoints( edgesImageData, 50, params.accuracy );
-		var edgeVertices = getVerticesFromPoints( edgePoints, params.vertexCount, params.accuracy, imageSize.width, imageSize.height );
-		var polygons = delaunay.triangulate( edgeVertices );
+		let blurredImageData = stackBlur.imageDataRGBA( tmpImageData, 0, 0, imageSize.width, imageSize.height, params.blur );
+		let greyscaleImageData = greyscale( blurredImageData );
+		let edgesImageData = detectEdges( greyscaleImageData );
+		let edgePoints = getEdgePoints( edgesImageData, 50, params.accuracy );
+		let edgeVertices = getVerticesFromPoints( edgePoints, params.vertexCount, params.accuracy, imageSize.width, imageSize.height );
+		let polygons = delaunay.triangulate( edgeVertices );
 
 		return addColorToPolygons( polygons, colorImageData, params );
 	} else {
